README
All three .py files must be in the same directory during the implementation. The naive Bayes model is a naive Bayes model made by me for explanatory purposes and to understand the model. This model, in most of the cases, has the same performance of the naive model implemented in scikit-learn.